+++ saleor/graphql/warehouse/schema.py
@@ -1,5 +1,6 @@
 import graphene
 import graphene_django_optimizer as gql_optimizer
+from django.db.models import QuerySet
 
 from ...core.permissions import ProductPermissions
 from ...warehouse import models
@@ -15,7 +16,18 @@ from .mutations import (
     WarehouseDelete,
     WarehouseUpdate,
 )
+from ..utils import sort_queryset
+from .sorters import (
+    WarehouseSortField,
+    WarehouseSortingInput,
+)
 from .types import Stock, Warehouse
+
+
+def sort_warehouses(qs: QuerySet, sort_by: WarehouseSortingInput) -> QuerySet:
+    if sort_by:
+        return sort_queryset(qs, sort_by, WarehouseSortField)
+    return qs.order_by("name")
 
 
 class WarehouseQueries(graphene.ObjectType):
@@ -27,7 +39,10 @@ class WarehouseQueries(graphene.ObjectType):
         ),
     )
     warehouses = FilterInputConnectionField(
-        Warehouse, description="List of warehouses.", filter=WarehouseFilterInput()
+        Warehouse,
+        description="List of warehouses.",
+        filter=WarehouseFilterInput(),
+        sort_by=WarehouseSortingInput(),
     )
 
     @permission_required(ProductPermissions.MANAGE_PRODUCTS)
@@ -37,8 +52,9 @@ class WarehouseQueries(graphene.ObjectType):
         return warehouse
 
     @permission_required(ProductPermissions.MANAGE_PRODUCTS)
-    def resolve_warehouses(self, info, **kwargs):
+    def resolve_warehouses(self, info, sort_by=None, **kwargs):
         qs = models.Warehouse.objects.all()
+        qs = sort_warehouses(qs, sort_by)
         return gql_optimizer.query(qs, info)